Refactored physical structure of proxy's security testing folder
