+++ spring-boot/proxy/src/test/java/eu/coatrack/proxy/security/ApiKeyFetcher_ApiKeyValueOfNullTest.java
@@ -1,4 +1,4 @@
-package eu.coatrack.proxy.security.api_key_fetcher_tests;
+package eu.coatrack.proxy.security;
 
 /*-
  * #%L
@@ -20,12 +20,11 @@ package eu.coatrack.proxy.security.api_key_fetcher_tests;
  * #L%
  */
 
-import eu.coatrack.proxy.security.exceptions.ApiKeyFetchingFailedException;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-public class NullApiKeyValueTest extends AbstractApiKeyFetcherTestSetup {
+public class ApiKeyFetcher_ApiKeyValueOfNullTest extends ApiKeyFetcher_AbstractTestSetup {
 
     @Test
     public void nullApiKeyValueShouldCauseException() {
